+++ benchopt/plotting/html/static/result.js
@@ -655,7 +655,6 @@ const makeLegend = () => {
     }
 
     let payload = {
-      title: solver,
       description: solversDescription[solver],
     }
 
@@ -744,7 +743,7 @@ const createLegendItem = (solver, color, symbolNumber) => {
 }
 
 
-function createSolverDescription(legendItem, {title, description}) {
+function createSolverDescription(legendItem, { description }) {
   if (description === null || description === undefined || description === "")
     description = "No description provided";
 
@@ -752,8 +751,7 @@ function createSolverDescription(legendItem, {title, description}) {
   descriptionContainer.setAttribute("class", "solver-description-container")
 
   descriptionContainer.innerHTML = `
-  <div class="solver-description-content">
-    <span class="solver-description-title">${title}</span>
+  <div class="solver-description-content text-sm">
     <span class="solver-description-body">${description}</span>
   </div>
   `;