+++ benchopt/tests/test_runner.py
@@ -55,18 +55,6 @@ def test_get_one_solution():
     expected = np.zeros(objective.X.shape[1])
     assert all(one_solution == expected)
 
-    # XXX - Remove in version 1.3
-    dataset = TEST_DATASET.get_instance(deprecated_return=True)
-    objective = TEST_OBJECTIVE.get_instance(deprecated_dataset=True)
-
-    with pytest.warns(FutureWarning, match="`get_data` should return a dict"):
-        objective.set_dataset(dataset)
-
-    with pytest.warns(FutureWarning, match="Objective should have a method"):
-        one_solution = objective.get_one_solution()
-    expected = np.zeros(objective.X.shape[1])
-    assert all(one_solution == expected)
-
 
 def _assert_parameters_equal(instance, parameters):
     for key, val in parameters.items():