+++ benchopt/utils/dependencies_mixin.py
@@ -101,8 +101,26 @@ class DependenciesMixin:
             try:
                 cls._pre_install_hook(env_name=env_name)
                 if cls.install_cmd == 'conda':
-                    install_in_conda_env(*cls.requirements, env_name=env_name,
-                                         force=force)
+                    if hasattr(cls, "requirements"):
+                        install_in_conda_env(*cls.requirements,
+                                             env_name=env_name,
+                                             force=force)
+                    else:
+                        # get details of class
+                        cls_type = cls.__base__.__name__.replace("Base", "")
+
+                        raise AttributeError(
+                            f"Could not find dependencies for {cls.name} "
+                            f"{cls_type} while it is not importable. This is "
+                            "probably due to missing dependency specification."
+                            " The dependencies should be specified in class "
+                            "attribute `requirements`.\n"
+                            "Examples:\n"
+                            "   requirements = ['pkg'] # conda package `pkg`\n"
+                            "   requirements = ['chan:pkg'] # package `pkg` in"
+                            "conda channel `chan`\n"
+                            "   requirements = ['pip:pkg'] # pip package `pkg`"
+                        )
                 elif cls.install_cmd == 'shell':
                     install_file = (
                         cls._module_filename.parents[1] / 'install_scripts' /
@@ -153,6 +171,21 @@ class DependenciesMixin:
             cls._pre_install_hook(env_name=env_name)
             if cls.install_cmd == 'conda':
                 conda_reqs = getattr(cls, "requirements", [])
+                if not is_installed and len(conda_reqs) == 0:
+                    # get details of class
+                    cls_type = cls.__base__.__name__.replace("Base", "")
+                    raise AttributeError(
+                        f"Could not find dependencies for {cls.name} "
+                        f"{cls_type} while it is not importable. This is "
+                        "probably due to missing dependency specification. "
+                        "The dependencies should be specified in class "
+                        "attribute `requirements`.\n"
+                        "Examples:\n"
+                        "   requirements = ['pkg'] # conda package `pkg`\n"
+                        "   requirements = ['chan:pkg'] # package `pkg` in"
+                        "conda channel `chan`\n"
+                        "   requirements = ['pip:pkg'] # PyPi package `pkg`"
+                    )
             elif cls.install_cmd == 'shell':
                 shell_install_scripts = [
                     cls._module_filename.parents[1] / 'install_scripts' /